Remove fetch calls and add lambdas to delay evaluation
Since capistrano/capistrano@f040f12353c3ad8f682631463743691b785b3353 the
order of evaluation of the configuration variables has changed. The
load:defaults task is evaluated before the user configuration, so there
is no need to fetch the variable before setting the default value.

The evaluation of the variables that depend on other variables should be
delayed using lambdas so if the user changes the base variable the
variables that depend on it get sane defaults.