Add an empty .rb file so Bundler.require doesn't break
Some applications use Bundler.require to require groups of gems.  When
the capistrano-torquebox is in one of these groups, Bundler will try to
load torquebox.rb which in turn loads load.rake and deploy.rake. These
files contain rake tasks and in this case they are getting loaded in the
context of a regular object not in the context of Rake. This will make
the Bundler.require fail with an error like this:

```
NoMethodError: undefined method `namespace' for main:Object
```

By adding this empty file, Bundler will load it instead of the
torquebox.rb and proceed happily.

The user of capistrano-torquebox could add the `require: false` option to
the gem entry in the Gemfile, but adding the file is a more elegant
solution and it is used by various "capistrano" gems.